Scaffolding: create base class, arduino and RPI derived classes
